pyproject.toml
[tool.poetry]
name = "searchy"
version = "0.1.0"
description = "A voice assistant based on phone calls"
authors = ["Juan Almanza <hi@scidroid.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7"
fastapi = "^0.81.0"
twilio = "^7.13.0"
python-dotenv = "^0.21.0"
wikipedia = "^1.4.0"
uvicorn = "^0.18.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
